fix: change the functionality of the application after the connection is re-established;
Change the current functionality of the application. It should not remove crdt-operations immediately after the re-connection.
Why? Because if the connection is lost again before user clicked "GET", then the value in offline mode is going to be calculated wrong.
Instead of that, operations should be removed once the user clicked to receive the latest values.